collect fixed-point iteration code into fixed-point module
